requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-http==5.2.2
apache-airflow-providers-amazon==9.7.0
minio==7.2.15
apache-airflow-providers-docker==4.4.0
apache-airflow-providers-postgres==6.2.0
requests
astro-sdk-python==1.8.1
pandas
apache-airflow-providers-discord==3.10.0